backend/scoring_service/requirements.txt
fastapi==0.110.0
uvicorn==0.23.2
python-multipart==0.0.6
numpy==1.23.5
scipy==1.10.1
librosa==0.10.1
soundfile==0.12.1
scikit-learn==1.3.2
joblib==1.3.2
torch
openai-whisper
